PyTest/AB_Soft_Automation_Task/requirements.txt
selenium
pytest
requests
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
oauth2client
httplib2
allure-pytest
allure-python-commons